requirements.txt
appdirs==1.4.4
asgiref==3.3.0
astroid==2.4.2
Babel==2.9.0
certifi==2020.6.20
colorama==0.4.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.5.0
django-environ==0.4.5
django-extra-views==0.13.0
django-haystack==3.0
django-heroku==0.3.1
django-oscar==2.1
django-oscar-api==2.1.0
django-phonenumber-field==3.0.1
django-psycopg2-extension==0.0.5
django-request-id==1.0.0
django-tables2==2.2.1
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djongo==1.3.3
dnspython==2.0.0
factory-boy==2.12.0
Faker==4.14.2
filelock==3.0.12
gunicorn==20.0.4
isort==5.6.4
jsonfield==3.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.2
passlib==1.7.4
phonenumbers==8.12.13
Pillow==8.0.1
pipenv==2020.8.13
portalocker==2.0.0
psycopg2==2.8.6
purl==1.5
PyJWT==1.7.1
pylint==2.6.0
pymongo==3.11.0
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.2.4
text-unidecode==1.3
toml==0.10.2
virtualenv==20.1.0
virtualenv-clone==0.5.4
whitenoise==5.2.0
wrapt==1.12.1
